full update

+++ api/v1/general/utils.py
@@ -19,9 +19,12 @@ def delete_object_file_post_delete(instance, field_name):
 
 
 def delete_object_file_pre_save(model_class, instance, field_name):
-    old_file = getattr(model_class.objects.get(id=instance.id), field_name)
+    try:
+        old_file = getattr(model_class.objects.get(pk=instance.pk), field_name)
+    except model_class.DoesNotExist:
+        return
     instance_file = getattr(instance, field_name)
-    if instance_file and instance_file != old_file and os.path.isfile(old_file.path):
+    if instance_file and old_file and instance_file != old_file and os.path.isfile(old_file.path):
         os.remove(old_file.path)
